% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{build.edgelist}
\alias{build.edgelist}
\title{For a sample, produce the transmission tree as a \code{igraph} object}
\usage{
build.edgelist(generator, sample)
}
\arguments{
\item{generator}{A list of class \code{tt.generator} produced by \code{tt.generator}.}

\item{sample}{A list of class \code{tt} produced by \code{sample.tt} or \code{sample.partial.tt}}
}
\value{
An \code{igraph} object
}
\description{
For a sample, produce the transmission tree as a \code{igraph} object
}
\examples{
generator <- tt.generator(stratus.example.tree)
samples <- sample.tt(generator, 1)
build.edgelist(generator, samples[[1]])
}
