% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_for_docs.R
\docType{data}
\name{grouping.map}
\alias{grouping.map}
\title{A vector assigning the tips of \code{stratus.example.tree} to groups (in the order they appear in \code{stratus.example.tree$tip.label}), as an example of multiple sampling.}
\format{A character vector of length 20}
\description{
A vector assigning the tips of \code{stratus.example.tree} to groups (in the order they appear in \code{stratus.example.tree$tip.label}), as an example of multiple sampling.
}
